requirements.txt
async-timeout==5.0.1
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
bs4==0.0.2
cachelib==0.13.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
Deprecated==1.2.17
Flask==3.1.0
Flask-Limiter==3.10.1
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
limiter==0.1.2
limits==4.0.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.19.0
ordered-set==4.1.0
packaging==24.2
passlib==1.7.4
pillow==11.1.0
pycparser==2.22
Pygments==2.19.1
pyotp==2.9.0
python-dotenv==1.0.1
qrcode==8.0
redis==5.2.1
requests==2.32.3
rich==13.9.4
soupsieve==2.6
SQLAlchemy==2.0.37
token-bucket==0.2.0
typing_extensions==4.12.2
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
zipp==3.21.0
